Integrate Flat Panel into Flatwizard

* The $$FRAMENR$$ placeholder is now saving the number as 4 digits

* found the flipflat on COM3

* basic commands almost working

* status and door working

* commands working

* added first unit tests and a bit of mock scaffolding

* adding more unit tests

* Unit test for driver are now decent and passing, C#-pified a bunch of code too

* fixed the enum case recommendation

* started working on integration, changed structure to fit

* seaprated sdk from model and went with command and response objects

* updated tests to reflect new connect

* view is displaying and connecting

* added unit tests now that I have coverage working

* can open and close the cover from the settings dialog

* setup dialog is working

* Looks to be working, need a image for the flat device though

* found the flipflat on COM3

* basic commands almost working

* status and door working

* commands working

* added first unit tests and a bit of mock scaffolding

* adding more unit tests

* Unit test for driver are now decent and passing, C#-pified a bunch of code too

* fixed the enum case recommendation

* started working on integration, changed structure to fit

* seaprated sdk from model and went with command and response objects

* updated tests to reflect new connect

* view is displaying and connecting

* added unit tests now that I have coverage working

* can open and close the cover from the settings dialog

* setup dialog is working

* Looks to be working, need a image for the flat device though

* removing Roslyn dependencies

* renamed NotOpenClosed to NeitherOpenNorClosed

* renamed IsOoo to EndsInOOO

* moved commands and responses to their own files and improved exception logging for sendcommand

* removed roslyn code analyzer stuff

* updated locale files with locale manager

* updated exception handling

* added negative unit tests, removed async from SendCommand to avoid thread conflicts

* made opening and closing the cover safer; added simulator device

* added interface for testability and unit tests for the FlatDeviceVM

* added slider for brightness and made setting brightness safer

* updated unit test to expect the new setBrightness bahavior

* added unit tests for alnitak simulator

* undoing the double-add for the FlatDevice - loving the icon for it :)

* added a setup dialog to the Alnitak devices, so the automatic discovery does not mess up other devices

* comitting so I can merge in upstream

* progress on auto detec and making it user friendly.

* now properly serializes the setup values

* autodetection and percentage slider are now working properly

* removed setup view for simulator and removing UI changes from Flatwizard

* all tests are passing, fixed a few bugs on disconnect and brightness handling.

* updated tests and view managers, corrected a localization mistake for german

* updated locale files with missing labels

* exposed the functions instead of the commands to make it easier to Mock for unit testing

* refactored tests to verify flat panel behavior


Former-commit-id: a5c0cfd6e565e0225c8f61f56b435dc1bf4c20dd
Former-commit-id: 3bd0149637df18bd52a137dd0cf7d1647280b1bd

+++ NINA/Model/MyFlatDevice/IFlatDevice.cs
@@ -16,7 +16,7 @@ namespace NINA.Model.MyFlatDevice {
 
         bool LightOn { get; set; }
 
-        int Brightness { get; set; }
+        double Brightness { get; set; }
 
         string PortName { get; set; }
